Rotation fixes
The right angle rotation algorithm has changed. Now, it uses only two rows in calculations instead of three, excluding the 3rd one. The rotated order orientation volume algorithms has also been changed. It gives priorities to values over 0.5f. It assumes that as soon as the values go below 0.5f, volumeRendering would give priorities to the negative below -0.5f values and snip the image to the according axis. The last statement is still questionable and will be resolved as soon as more data is aquired and tested